#148 test: 존재하지 않는 여행 일정 삭제를 시도하면 예외가 발생하는지 테스트

+++ backend/src/main/java/moheng/planner/application/PlannerService.java
@@ -73,6 +73,10 @@ public class PlannerService {
 
     @Transactional
     public void removeTripSchedule(final Long tripScheduleId) {
+        if(!tripScheduleRepository.existsById(tripScheduleId)) {
+            throw new NoExistTripScheduleException("존재하지 않는 여행 일정입니다.");
+        }
+
         tripScheduleRepository.deleteById(tripScheduleId);
     }
 }
